Moved generators into their own package
